requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.8
django-model-utils==4.1.1
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==9.2.0
ipython==7.28.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.0
vine==5.0.0
wcwidth==0.2.5
